pyproject.toml
[project]
name = "my-oci-toolkit"
version = "1.0.1"
description = "A personal, intelligent wrapper for the OCI CLI."
authors = [{ name = "Veno Ajie", email = "ven.ajie@protonmail.com" }]
requires-python = ">=3.8"
dependencies = [
    "typer[all]",
    "python-dotenv",
    "rich",
    "rapidfuzz",
    "PyYAML",  
    "jsonschema", 
]

[project.scripts]
my_oci = "my_oci.cli:app"